Create ShooterSolenoid with Methods & Create ShooterHood Command
Created a shooter solenoid in the pneumatics subsystem as well as the setShooterForward() and setShooterReverse() methods to use with it (shoot forwards or backwards). The MoveShooterHoodCommand was also created so that the shooter hood could be moved and tested so that it could be changed as needed. Some motor ID constants were also adjusted and some constants were created as well as necessary to do all of this.