Create TwoBallAuton, Tune PID Constants, Update Constraints, & Modify ThreeBallAuton
Created the TwoBallAuton command and tuned/fixed the PID constants so that the robot could move and shoot with this and other commands effectively at both a slow and fast speed. To do so, the constraints were also changed. Additionally, the other autons were updated to reflect these changes and the ThreeBallFromMiddleTest was renamed to ThreeBallAuton and modified a bit as well to work with the fixed PID constants and constraints.