Create Limit Switches, Invert Limit Switches, & Fix Prac Bot MAC Address Issue
Created the code used for the limit switches as well as the code to grab onto and release the climb bar. In doing so, the existing code to help do so from the GrabBarCommand and ReleaseBarCommand was put into the MoveClimberCommand file, which was modified as needed to make this work based on the new code and information from the Phoenix documentation. Limit switches were also inverted and methods were created to get them as well as check if they need to be created/are on the robot. The constants, robot container, and other classes were all appropriately updated in the process. The issue with the Prac Bot MAC Address in the robot container was also fixed.