Fix Climber Motor Issues, Add Equation, & Add Controller Rumble
Fixed issues with the climber motors (arms) not going the same speed and going in different directions (due to inversion). Also implemented the equation into the DriveCommand and implemented controller rumble so that, when the robot does not see a target, the controller will rumble to indicate that vision tracking should not be used. Otherwise, it will not rumble/stop current rumble.