Fix New Climber Motor Speed & Shutoff Issues
Fixed the code that syncs up the positions and made it check for both if the arms are going up and down. This way, the correct climber arm would be shut off in all cases as opposed to only when the climber arms are rising up. Now, the code checks to see which direction the climber is going, and then looks to see if one arm is 500 or more units ahead of the other. If so, it disables that arm until the other arm has caught up with it (difference between the two is less than 500 units).