Create LightsSubsystem & Add CANdle Basic Code
Created the LightsSubsystem for the candle/lights and added the basic code and constants needed to control these lights through the CANdle. To make this work, more code must be put in later. However, this added code serves as a starting point to add more to for this. A LightsCommand will also need to be created later that has not been yet.