Work on Climber Issue + Lights Code & Fix Errors
Fixed a wide variety of errors/issues in the code that needed to get resolved and worked on the climber issue. It should work now, but more testing needs to be done to confirm this. The code for the robot lights was also adjusted a bit and other errors encountered during testing, such as not handling a null pointer that could rarely occur when getting the vision tracking yaw, were fixed as well. To do all o this, the constants were added to, changed, and updated as needed as well as some of the other files as well.