Add to Vision Tracking & Update Constants
Added to vision tracking the ability to determine if the shooter hood should be used when shooting as well as updated/added linear regression equations to determine what the motor speeds need to be at given where the robot is at to shoot. Both auton files, the constants files, and other files and commands were updated to do all of this and update gear ratio and steer reduction constants as necessary.