Add Color Sensor Methods & Constants
Added various methods for the color sensor to check if the robot has a ball and, if so, what color the ball is. To do so, various constants were created to do so as needed. However, this may still need to be added to as the hasRedBall() method may need to be used in the hasBall() method if the program determines that the robot has a ball.

Furthermore, a climbing issue was fixed as it was being tested so that the practice bot can now move its arms up and down to climb again. Additionally, the code for the candle was updated slightly so that issues with the lights could be successfully diagnosed and be prepared for future testing (setting the lights to yellow should be changed back later as needed).