Add ZeroGyroCommand & Fix knockDownMotorSpeed
Added the ZeroGyroCommand (to reset the gyroscope) to the code and fixed the speed of the knockdown motor by multiplying it by the gear ratio of the knockdown belts. Other files were adjusted as needed and other fixes with the steer ratio and the drive ratio were implemented as well.