Create Override for Climbers
Created an override feature for the climbers so that, if they do not start in the same position, the code that will try to sync them up using the encoders (and will not do so because they are in different positions) will not be run while up in the d-pad is hit to activate the override. The left soft bottom and right soft bottom getter methods were also added and a little bit of testing was done with the lights.