Create & Add to Soft Bottom Limits + Override
Created the soft bottom limits for the climber arms by manually looking at the constants. Furthermore, we began to add more code to these and the override that will continue to be worked on. To do so, the constants in the ClimberSubsystem were added to as needed the climber and pneumatics subsystems had code added to them as needed.