Automate Climber & Add SM2 Code
Automated the process of climbing (not speed, also added code to ensure that it can be manually overridden for the regular manual climb if needed) and added the SM2 code. Also adjusted/added to the constants, robot container, and other subsystems as needed to do so. Motor speeds were also adjusted as needed from testing from the night before, but more testing still needs to be completed.

Changes from the previous night are also uploaded here, with some smaller changes being made to the code for the lights to help figure out the issue with the lights being purple sometimes when they are not supposed to be.