Pittsburgh Comp Updates & Swap Shooter Camera
Consists of the various changes made during Pittsburgh (primarily adjusting values to ensure that shooting, vision tracking, and the autons worked better). During this time, a new auton was created for the left tarmac as well. In the process, various things were updated, including some of the comp constants. Particularly, the old shooter camera (at the bottom) was swapped with the new one, with the constants being modified to reflect that.