Modify for Citrus & Simplify Climber
Modified the code as needed for the citrus climber to work and simplified it to make it less automatic, but also easier to use for testing purposes on the practice bot. Created on a separate branch to make it easier to test while keeping it separated from the main code that is used for the robot.