Add Climber code, limits, & right inversion
Added the necessary climber code to make it so that the citrus climber would work. In doing so, limits were added to both the left and the right climber arms, the right arm was un-inverted, and various other changes were made to the constants and other methods were added to ensure that the climber would work.