Add Moving Shot Capabilities
Added to the robot the ability to shoot while moving in another direction using vision tracking to account for changes in the trajectory of the ball from movement. To do so, various methods were created and various other methods and constants had their code modified, modifying things like linear regression equations as needed.

While this does work when moving slowly side to side and somewhat works with the first shot when moving forward slowly, this still needs work to ensure that it functions in all situations. Adjusting some of the values of the equations or other variables might do this, but more testing is needed and will be done.