Four/Five Ball Auton & Macomb Comp Updates
Updated the four/five ball auton to ensure that it works (although it still needs some work to make sure that it is more consistent) and updated other code in other files as needed a bit before and during the comp. This was done for the climber and the shooter command to ensure that shooting would work appropriately and as needed when certain conditions were met (revving up and actually shooting).