Modify Shooting Variables & Implement Ball Tracking
Modified the code so that various variables in shooting that were making it more difficult to shoot (front camera PID and tolerance) were adjusted. In addition, code was added to turn on ball tracking when the left trigger on the driver controller is pressed. In doing so, the driver mode of the camera can be turned on and vision tracking can be used to guide the robot to cargo to make it easier to intake into the robot. More needs to be done to ensure that other things with vision work, but will be done at a later time.