Modify & Fix Auton Code & Create Quicker Four/Five Ball Auton
Modified and fixed various issues with the auton code to allow the autons to have various other features available to them. Specifically, the ability to use vision tracking and the ability to use the cameras for ball tracking. To do so, the swerve controller and trajectory command generator files had their code modified in various ways. After being modified/added to, this code was then fixed to ensure that the right variables were used and not being confused with other things. All of the autons were also updated to reflect these changes (as more variables were added to the parameters of constructors used for the auton commands).

Furthermore, a new, quicker four ball auton was created that is still in the process of being worked on to ensure that it fully works (can pick up the fourth and fifth ball) and is consistent once fine tuning is finished.

At the moment, the code still needs to be modified to ensure that ball tracking works for the blue balls as well and not just for the red. All of the other modifictions that needed to be made to the original FourBallAuton have, as far as I know, been made.