Create and Fix Quicker Four Ball Auton
Created and fixed the quicker four ball auton so that it would work for all of the balls. In doing so, the remainder of the path for it was created (with the ball from the human player being picked up before the ball at the terminal as this was changed from the original plan). Additionally, this was fine tuned so that it would work appropriately.

Side note: In the last commit, blue ball tracking was created. However, I did not realize this and said it was not.