Tune PID Constants & Shooter RPMs for Q. Four Ball
Tuned the PID constants and the shooter RPMs for the four/five ball auton a little. The PID constants were not changed yet as we decided to revert back to the original values and the shooter RPMs were changed slightly to make things much more consistent. In order to make sure that each point is being made by the robot, however, much more tuning will need to be done with the PID constants.